requirements.txt
newspaper3k
nltk
fastapi
uvicorn[standard]
lxml_html_clean